+++ package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "type": "commonjs",
   "scripts": {
-    "test": "jest --verbose"
+    "test": "jest --verbose",
+    "type": "module"
   },
   "repository": {
     "type": "git",
@@ -18,20 +19,25 @@
   },
   "homepage": "https://github.com/slugcraft/TECH-NEBRIOS-BACKEND#readme",
   "dependencies": {
+    "@prisma/client": "^6.7.0",
     "aws-sdk": "^2.1692.0",
     "bcryptjs": "^3.0.2",
     "body-parser": "^2.2.0",
+    "date-fns": "^4.1.0",
     "dotenv": "^16.4.7",
     "ejs": "^3.1.10",
+    "exceljs": "^4.4.0",
     "express": "^5.1.0",
     "express-session": "^1.18.1",
-    "jsdoc": "^4.0.4",
+    "format": "^0.2.2",
     "jsonwebtoken": "^9.0.2",
     "mariadb": "^3.4.0",
     "mysql2": "^3.14.0",
     "nodejs": "^0.0.0"
   },
   "devDependencies": {
-    "jest": "^29.7.0"
+    "jest": "^29.7.0",
+    "jsdoc": "^4.0.4",
+    "prisma": "^6.7.0"
   }
 }
