+++ server.js
@@ -11,13 +11,13 @@ const app = express();
 
 /* Rutas */
 const charolaRoutes = require("./routes/charola.routes");
-const comidaRoutes = require("./routes/comida.routes");
+const alimentoRoutes = require("./routes/alimento.routes");
 const hidratacionRoutes = require("./routes/hidratacion.routes");
 /* ----- */
 
 /* Rutas de la API */
 app.use("/charola", charolaRoutes);
-app.use("/comida", comidaRoutes);
+app.use("/alimentacion", alimentoRoutes);
 app.use("/hidratacion", hidratacionRoutes);
 /* ----- */
